Mailing failed because of non-escaped slashes
When trying to mail if one of the arguments contains a slash an empty mail would be send. This happens mostly if a postprocessing script is set.

The fix is to show the filename only (not the full path). I couldn't get the slashes to escape properly. As the escaped string would have to be escaped as well before it's used and after a while I lost track of the amount of regular slashes and escaped slashes to use.