Added remote_path option to CouchPotato section, allows setting the basepath for where CouchPotato looks for its video files on the server its installed on.
Re-coded autoProcessMovies, it now will attempt to get imdbID from dirName, nzbName and if both fail it then uses a regex to get the movie title and uses that to perform a search on CouchPotato.

We also now get the release_id and use that instead of movie id.

Fixed a bug that was preventing failed downloads from working properly for CouchPotato.