Report Downloader failures to SickRage
When extraction is done by SickRage, report failures from Downloader to SickRage.
For example, when download fails in SAB because out of retention SAB reports failed status -1. Or when verification fails SAB reports failed status 1. In such cases also report failure to SickRage. Without this fix, this results in "Returned log from SickBeard was not as expected.!"